refactor(atlas): lift country badge rule to shared catalog_badges (#947)
badge_for() moved verbatim (all three branches: US hardcoded RICH, module
countries via catalog_status, no-module default SAMPLE) so the all-regions
atlas and the Explorer feed share one source. Coverage CSV regenerated
byte-identical (identity gate in tests).
